+++ src/main/java/com/example/d3nserver/auth/authToken/request/AuthRequest.java
@@ -1,4 +1,4 @@
-package com.example.d3nserver.auth.common.request;
+package com.example.d3nserver.auth.authToken.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
